pyproject.toml
[tool.poetry]
name = "danbi-task-python"
version = "0.1.0"
description = ""
authors = ["hanihoneykim <107341819+hanihoneykim@users.noreply.github.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
django = "3.2"
djangorestframework = "3.11.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
